requirements.txt
asgiref==3.8.1
Django==5.1.7
dotenv==0.9.9
pdf2image==1.17.0
pillow==11.1.0
psycopg==3.2.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn==23.0.0
whitenoise==6.8.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
dj-database-url==2.3.0
PyMuPDF==1.24.14
django-resized==1.0.2
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-import-export==4.0.0